requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
--extra-index-url https://download.pytorch.org/whl/cu118
diffusers==0.32.2
transformers==4.48.0
numpy==1.26.4
accelerate==0.28.0
librosa==0.11.0
soundfile==0.12.1
opencv-python==4.9.0.80
einops==0.8.1
gradio==5.24.0
huggingface_hub==0.30.2
omegaconf==2.3.0
ffmpeg-python==0.2.0
imageio==2.31.1
gdown
requests
imageio[ffmpeg]

# GPEN-BFR Face Enhancement Dependencies (ONNX-based)
onnx>=1.17.0
onnxruntime-gpu>=1.22.0  # GPU acceleration for GPEN-BFR enhancement
scipy>=1.15.0
psutil>=7.0.0

# Fallback for systems without CUDA
# If onnxruntime-gpu fails, manually install: pip install onnxruntime>=1.22.0

# GFPGAN Face Enhancement Dependencies
gfpgan>=1.3.8
basicsr>=1.4.2
facexlib>=0.3.0
realesrgan>=0.3.0
yapf>=0.40.1